requirements.txt
langgraph>=0.2.50
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-mcp-adapters>=0.1.0
mcp>=1.0.0
yfinance>=0.2.40
pandas>=2.2.0
lxml>=5.0.0
httpx>=0.27.0
google-api-python-client>=2.140.0
google-auth>=2.34.0
google-auth-oauthlib>=1.2.0
